# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Skincare Seller",
        "description": "Laravel project to practice developing Rest API with L5 Swagger OpenApi",
        "contact": {
            "name": "Willy",
            "email": "willy.422023024@civitas.ukrida.ac.id"
        },
        "version": "1.0",
        "x-logo": {
            "url": "https://tse3.mm.bing.net/th?id=OIP.LsfqL0yFaOPMLTKynENKEgHaF7&pid=Api&P=0&h=180"
        }
    },
    "paths": {
        "/api/user/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register new user & get token",
                "operationId": "register",
                "requestBody": {
                    "description": "Request body description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            },
                            "example": {
                                "name": "willy rusli ukrida",
                                "email": "willyukrida@gmail.com",
                                "password": "willy123ukrida",
                                "password_confirmation": "willy123ukrida"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "201": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/user/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Log in to existing user & get token",
                "operationId": "login",
                "requestBody": {
                    "description": "Request body description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            },
                            "example": {
                                "email": "willyukrida@gmail.com",
                                "password": "willy123ukrida"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "201": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/user/logout": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Log out & destroy self token",
                "operationId": "logout",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "User Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "201": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport_token_ready": [],
                        "passport": []
                    }
                ]
            }
        },
        "/api/skincare": {
            "get": {
                "tags": [
                    "skincare"
                ],
                "summary": "Display a listing of items",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "skincare"
                ],
                "summary": "Store a newly created item",
                "operationId": "store",
                "requestBody": {
                    "description": "Request body description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Skincare"
                            },
                            "example": {
                                "nama": "Cetaphil Gentle Skin Cleanser",
                                "brand": "Cetaphil",
                                "category": "Cleanser (Pembersih Wajah)",
                                "price": "150000",
                                "stock": "1000",
                                "description": "Cetaphil Gentle Skin Cleanser adalah pembersih wajah yang diformulasikan untuk kulit sensitif..",
                                "image": "https://tse4.mm.bing.net/th?id=OIP.4lndUWsMx42RQGJSicHXKAHaHa&pid=Api&P=0&h=180"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "201": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport_token_ready": [],
                        "passport": []
                    }
                ]
            }
        },
        "/api/skincare/{id}": {
            "get": {
                "tags": [
                    "skincare"
                ],
                "summary": "Display the specified item",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of item that needs to be displayed",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int65"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Item not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "skincare"
                ],
                "summary": "Update the specified item",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of item that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Request body description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Skincare"
                            },
                            "example": {
                                "nama": "Kiehl's Ultra Facial Cream",
                                "brand": "Kiehl's",
                                "category": "Moisturizer (Pelembap)",
                                "price": "200000",
                                "stock": "1500",
                                "description": "Pelembap yang memberikan hidrasi 24 jam dengan tekstur ringan dan cepat meresap.",
                                "image": "https://feelunique.com/cdn-cgi/image/quality=75,format=auto,metadata=none,dpr=1/img/products/61024/alternative/Kiehl__039_s_Ultra_Facial_Cream_50ml_1_1523698750.jpg"
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Item not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport_token_ready": [],
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "skincare"
                ],
                "summary": "Remove the specified item",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of item that needs to be removed",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Item not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport_token_ready": [],
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Skincare": {
                "title": "Skincare model",
                "description": "Skincare model",
                "required": [
                    "nama",
                    "brand"
                ],
                "xml": {
                    "name": "Skincare"
                }
            },
            "User": {
                "title": "Skincare model",
                "description": "Skincare model",
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "xml": {
                    "name": "Skincare"
                }
            }
        },
        "securitySchemes": {
            "passport_token_ready": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format without \"Bearer \"",
                "name": "Authorization",
                "in": "header"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}